# api/rotaract-bar.openapi.yaml
openapi: 3.0.3
info:
  title: OpenAPI definition
  version: v1
servers:
  - url: http://localhost:8080
    description: Generated server url
paths:
  /api/orders:
    get:
      tags:
        - purchase-order-controller
      summary: Get all orders
      description: This endpoint gets all currently saved orders
      operationId: getOrders
      responses:
        "200":
          description: Successful
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/PurchaseOrder"
        "500":
          description: Internal server error
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/PurchaseOrder"
        "400":
          description: Bad request
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/PurchaseOrder"
    post:
      tags:
        - purchase-order-controller
      summary: Create new Order
      description: This endpoint creates a new purchase order
      operationId: createOrder
      requestBody:
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/PurchaseOrder"
        required: true
      responses:
        "201":
          description: Successful
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/PurchaseOrder"
        "400":
          description: Bad request - Invalid details provided
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/PurchaseOrder"
        "500":
          description: Internal server error
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/PurchaseOrder"
  /api/products:
    get:
      tags:
        - product-controller
      operationId: getProducts
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Product"
  /api/ingredients:
    get:
      tags:
        - ingredient-controller
      operationId: getIngredients
      summary: "Get all ingredients"
      responses:
        '200':
          description: OK
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/Ingredient'
  /api/orders/batch:
    post:
      tags:
        - purchase-order-controller
      summary: Batch import orders
      description: Imports multiple purchase orders in a single request. Orders that fail validation are skipped and reported in the errors list.
      operationId: importOrdersBatch
      requestBody:
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/BatchOrderImport"
        required: true
      responses:
        "207":
          description: Multi-status — contains created orders and any import errors
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/BatchOrderResult"
        "400":
          description: Bad request - request body is invalid
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/BatchOrderResult"
        "500":
          description: Internal server error
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/BatchOrderResult"
  /api/orders/{id}:
    get:
      tags:
        - purchase-order-controller
      summary: Get Order by id
      description: This endpoint gets all orders by provided id
      operationId: getOrderById
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
            format: int64
      responses:
        "200":
          description: Successful
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/PurchaseOrder"
        "400":
          description: Bad request - Invalid id provided
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/PurchaseOrder"
        "500":
          description: Internal server error
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/PurchaseOrder"
components:
  schemas:
    OrderedItem:
      type: object
      properties:
        id:
          type: integer
          format: int64
        productId:
          type: integer
          format: int64
        quantity:
          type: integer
          format: int32
          minimum: 1
        bottleSale:
          type: boolean
          default: false
          description: "Marks this item as a bottle sale for analytics"
        customPrice:
          type: number
          format: double
          minimum: 0
          description: "Negotiated price for bottle sales; overrides quantity × product price in totals"
    Product:
      type: object
      properties:
        id:
          type: integer
          format: int64
        name:
          type: string
        price:
          type: number
          format: double
        category:
          type: string
          enum:
            - DRINKS
            - BEER_WINE_NONALC
            - SHOTS
          description: "Product category for bar organization"
        ingredientIds:
          type: array
          items:
            type: integer
            format: int64
          description: "List of ingredient IDs for this product"
      required:
        - id
        - name
        - price
        - category
    Ingredient:
      type: object
      properties:
        id:
          type: integer
          format: int64
        name:
          type: string
      required:
        - id
        - name
    PurchaseOrder:
      type: object
      properties:
        id:
          type: integer
          format: int64
        orderedAt:
          type: string
          format: date-time
        items:
          type: array
          items:
            $ref: "#/components/schemas/OrderedItem"
        returnedCupsCount:
          type: integer
          format: int32
          minimum: 0
        tipAmount:
          type: number
          format: double
          minimum: 0
          description: "Optional tip amount in EUR"
        paymentMethod:
          type: string
          enum: [ CASH, PAYPAL ]
          description: "Payment method used for this order"
    BatchOrderImport:
      type: object
      required:
        - orders
      properties:
        orders:
          type: array
          items:
            $ref: "#/components/schemas/PurchaseOrder"
          minItems: 1
    BatchOrderResult:
      type: object
      properties:
        created:
          type: array
          items:
            $ref: "#/components/schemas/PurchaseOrder"
          description: "Successfully imported orders"
        errors:
          type: array
          items:
            type: string
          description: "Error messages for orders that could not be imported"